Compose: refactored Shibboleth containers to avoid local build dependencies
* Shibboleth nginx config has been moved to am-shib
* Shibboleth idp and ldap are now part of shib-local

Bootstrap is now done as part of the container, not as part of the build. For
the most part, this is done in the 'command' by running a 'bootstrap.sh' script
prior to the intended command (e.g. run-jetty.sh). However, for ldap we make
use of a feature of the Osixia openldap image, which allows multiple processes
to be run at start up. We install our bootstrap script as the 'zz_bootstrap'
service, alongside the main 'slapd' process.

The shib-local services are optional and will eventually be replaced with an
external IdP.