deployment: change playbooks to tag images with specific version
We now tag the built Docker images with both "latest", as before, but also the specific version number too. This isn't supported by the docker_image Ansible module, so the playbooks have been changed to use the docker command directly instead.

The playbooks used to use 'repos' and 'images' lists to describe what we wanted to build. This was inefficient and unwieldy, since the version/tag was repeated and the link between image and repo were implicit.

With this change, we now use a 'projects' list, which includes both repo and image information. This means that images now take the version from the enclosing project, and there can be multiple images per project.

In addition, Travis CI is now configured to run ansible-lint and yamllint against the Ansible playbooks.