compose: add qa config based on dev config
Now that we are using the Ansible playbooks to build images for QA
deployment, we want Docker Compose to pull these built images rather than
attempting to (re-)build them itself. Because of the way that Compose
configurations work, if there is a build section for a service, the image
is always built, not pulled. We therefore need to create a separate config
file for dev and qa, to force images to be pulled for qa and built for dev.

I've therefore copied the existing dev config as 'qa', and trimmed down
the dev config to just include the build info for each service. The make
has been updated to accommodate that: it now uses qa by default and I've
added an additional ENV make parameter that can be used to override this,
i.e. ENV=dev to enable the dev build instead of qa.

The Docker Compose configuration now allows the specific version of each
of our main images to be given as environment variables, rather than just
assuming 'latest' (although this is still the default).

The registry address can also be given, allowing us to pull the images
from a specific private/custom registry instead of just using Compose to
build them, for example 'localhost:5000/'.