compose: don't wipe databases during bootstrap for qa containers
We currently run a DROP DATABASE and then a CREATE DATABASE every time we
"bootstrap" the Dashboard and Storage Service containers. This is okay for
a dev environment, but for QA we want these to survive multiple redeploys,
so the database should not be dropped as part of the deployment.

Also wait to be able to run an actual mysql query before attempting database
bootstrap, rather than an arbitary 30 seconds.