deployment: checkout QA copy of channel adapter to different folder
This is to avoid the case where the main images playbook has already been
executed and so the Git clone of the rdss-archivematica-channel-adapter
already exists. If this happens, the git-clone task in the QA playbook will
detect that nothing has changed, and the docker images for minikine and
dynalite will not be built or published, causing deployment to fail.

Doing a separate checkout for the QA playbook fixes this.