compose: fix incorrect path for msgcreator checkout in QA config
In a previous commit I changed the path that the msgcreator source gets
checked out to. I forgot that the path was used in the QA compose config,
as the path for a volume. Consequently the QA compose config broke if the
msgcreator checkout didn't still exist in the old location.