compose: fix playbooks to always use valid docker tags
When using a non-semver "version" like a branch name (e.g. 'qa/jisc')
the string being used to tag built Docker images wasn't valid syntax
according to the Docker tag format. This fixes that, replacing invalid
characters with '_' (currently only '/', since this is the most likely
to appear within a branch name). We also truncate the name string to 128,
to adhere to the limit imposed by Docker.