deployment: add AWS deployment scripts
This allows the resources required for deploying RDSSARK services to Amazon
Web Services to be created and updated in a consistent, methodical manner.

The deployment process is implemented in Bash, and uses the AWS CLI and
Docker Machine to enact the deployment. There are two scripts: deploy.sh,
which deploys the required resources, and teardown.sh, which removes them
all when no longer required.

The Ansible playbooks are used for building the Docker images, and the built
images are deployed to a local registry running on the Docker host that
Docker Machine provisions to.

As well as the container-based services, the deployment also includes EC2
instances for an Arkivum appliance and NFS server. In addition, Route53 is
used to provide public and private DNS services, which an external domain
can be configured to use.

Each deployment is configurable through the single configuration file and
environment variables. Amongst other things, configuration controls whether
or not an Arkivum appliance is included in the deployment, and whether or
not Shibboleth authentication is enabled, as well as instance and volume types
and volume sizes for each instance.

Resource tag are used to allow AWS resource to be easily identified as being
for a particular project and environment. The main resources that are tagged
are EC2 instances, EBS volumes and security groups. The primary use of these
tags is to allow tagged resources to be removed as part of teardown, instead
of being orphaned.