compose: add bootstrap for RDSS storage service locations
This adds Transfer Source Locations for "automated" and "interactive",
which are required for the automated and interactive workflows in RDSS.

The code is in a Python script that is mounted into the SS container,
and then executed from the Make bootstrap target. Though we'd rather avoid
adding a volume mount, this is much cleaner than trying to call python
from bash from Make, with all the escape characters that entails.