compose: idp now waits for metadata providers before starting
There is a race condition between the IdP starting and its metadata
providers (the SPs, i.e. Archivematica apps) starting. If the providers are
not up and running before the IdP attempts to contact them then the IdP
will fail to start.

We now therefore parse the metadata-providers.xml and wait for each
metadataURL mentioned to become available to the IdP.

This is necessary because the `sleep 30` which was arbitrarily part of the
bootstrap previously has been removed, unmasking this issue.