compose: fix version string handling
Use of VERSION in Makefiles was inconsistent - it was being passed to some
docker-compose commands but not all. This caused problems for those that
didn't get it, especially since there was no default (there is now). We now
export VERSION within the Makefile so that commands run within targets can
pick it up without it having to be specified explicitly for every command.

Also moved build args from qa (which shouldn't have a build section) to the
dev config (which should).