compose: fix errors in create SS storage locations bootstrap
Creating the storage locations in the SS would be part of the SS bootstrap,
right? Wrong: whilst this is the logical place for it, it requires a
pipeline to have been created, which means it's actually part of the
Dashboard bootstrap instead.

Also fixed small typo in the create locations script itself.