compose: change dev bootstrap to use same logic as qa bootstrap
Previously we were doing a DROP..CREATE on the databases during bootstrap.
This created inconsistent results: Archivematica doesn't behave well if its
database is lost but its files remain. Instead, we now do what we do in qa:
use CREATE IF NOT EXISTS to create the db, so that we don't clobber what is
already there.