compose: add destroy-volumes target
We have create-volumes but no way to clean them up after destroy. The
destroy-volumes target fills this gap, removing all of the docker volumes
we create and also removing all files from the defined paths. This is
obviously dangerous, so we ask the user to confirm their intent before
doing anything.

Note that the removal of the files will likely require sudo rights.