compose: remove use of ports in entity ids for Shibboleth
Previously we had two apps, Dashboard and Storage Service, running on
different ports (443, 8443) on the same host (archivematica.example.ac.uk).
To remove the ports from the entity ids for these apps, it was necessary to
change this so they run on two different vhosts on the same port (443). The
vhosts are as follows:

* Archivematica Dashboard = dashboard.archivematica.example.ac.uk
* Archivematica Storage Service = ss.archivematica.example.ac.uk

These hostnames are configurable via environment variables.

The IdP too was using the port in its entity id, so this too was changed.
This meant it was necessary to change its exposed port from 6443 to 4443,
which means that its internal port is the same as its exposed port.

The following environment variables have changed with this commit:

* NGINX_HOSTNAME is replaced by AM_DASHBOARD_HOST and
  AM_STORAGE_SERVICE_HOST
* AM_DASHBOARD_EXTERNAL_PORT and AM_STORAGE_SERVICE_EXTERNAL_PORT are
  replaced with AM_EXTERNAL_PORT

The use of different server names introduces the potential for very
long server names, so the nginx `server_names_hash_bucket_size` has
been increased to accomodate this.