shibboleth: add support for using custom certs
The Dashboard, Storage Service and IdP containers are now able to use
custom PKI keys and certificates instead of assuming that we want to
generate some. The `GENERATE_SSL_CERTS` Makefile variable may be used to
control whether they are generated or not. The default is to use generated
certificates.

If `GENERATE_SSL_CERTS` is set to `false`, other environment variables come
into play. The `AM_DASHBOARD_SSL_*` variables relate to the Dashboard,
`AM_STORAGE_SERVICE_SSL_* variables relate to the Storage Service, and the
`IDP_SSL_*` variables relate to the IdP. More details are given in the
readme documentation.

In the case of the IdP, the provided key and certificates are needed in a
number of different formats, so the bootstrap script (run as part of the
entrypoint) takes care of converting the input files into the required
formats and locations.