shibboleth: add scripts to create and use test PKI certificates
These scripts create a two-tier CA, with a root and intermediate, and the
required certificates for the AM Dashboard and Storage Service, and the
example IdP.

The two-tier-ca/init.sh script takes care of the CAs. The certs and keys
for the services are created by create-test-certs.sh (which also runs the
init script).

Lastly, the test-custom-certs.sh is a test harness that defines some
domains, creates certs and keys for them, brings up the docker env to use
Shibboleth, and then reconfigures it to use the test certificates. This
then allows a user to confirm that the AM services are available under the
defined test domains.