shibboleth: use signing certificates in metadata fetch
The MetadataProvider config in shibboleth2.xml has been updated to use a
MetadataFilter to verify that the fetched metadata is signed by the
certificate given by the `SHIBBOLETH_METADATA_SIGNING_CERT` environment
variable. If none is given, a default UKAMF signing cert is used.

Adding signature support meant re-enabling the security policy (previously
we were using NullSecurity). This meant providing the PKIX key and certs
in various places, through TrustEngine and CredentialResolver sections.
This is all working now - the SP successfully downloads metadata and
authenticates with the IdP.