deployment: use docker-compose v1.17.0 by default
This is because versions > 1.19.0 have a change in their behaviour
(introduced in 1.19.0) that means that our use of docker-compose to exec a
mysql command to test mysql connectivity during the bootstrap doesn't
return the right status, so the deployment gets stuck waiting for the
mysql service to become ready.

There is also a different problem with v1.18.0, which attempts to use a
newer version of the Docker API, which means that `docker-compose build`
fails, hence why v1.17.0 is used here.

Fixes issue #191.