add quickstart script to quickly and easily start/stop services for a user
As new people come to the project, so we need to get them up and running
quickly and easily. The quickstart script allows them to run simple
commands to start and shutdown a Docker compose environment.

It also makes provision for having multiple deployments running on the same
Docker machine, using the user's username to create a unique namespace for
each deployment and checking ports to ensure that only unused ones are
assigned to deployed services.

The deployed services use the default config: no Shibboleth is enabled, and
the 'qa' environment is used rather than 'dev'.

This is intended for development purposes only; it would take a bit more
work and hardening to make it suitable for qa/production deployments.