deployment: set channel adapter config during deployment
We now set various additional environment variables that configure the
channel adapter. This means we can now configure it to use real AWS
services, rather than assuming we want to use the mock services.

The MOCK_AWS env var controls whether or not the mock AWS services are
deployed. If set to false, the channel adapter will be configured to use
real AWS services using the settings provided, otherwise these settings
will be ignored.

If real services are to be used, these must be set up prior to deployment.
They will not be provisioned by the deployment script.