deployment: check AWS CLI version is sufficient
We're now using 'dynamodb tag-resource' command, which was introduced in
v1.11.80 of the AWS CLI. This is newer than the default version available
on Ubuntu 14.04 LTS, for example, so requires the pip version be used
instead.

If the installed version of AWS CLI is not sufficient, the deploy script
will now stop will a fatal error.