deployment: update teardown to include DynamoDB tables and Kinesis streams
We now teardown any DynamoDB tables that have been created. We also
optionally teardown the Kinesis streams too, but since these may be used by
other parts of the RDSS platform we do not do this by default. Instead, we
look to the DESTROY_KINESIS_QUEUES env var to see if we need to do so.