Deployment: configure and provision AWS resources for channel adapter (#215)
* compose: set IAM role for channel adapter to use

We now allow the IAM role ARN for the channel adapter to use when
connecting to Kinesis.

* deployment: set channel adapter config during deployment

We now set various additional environment variables that configure the
channel adapter. This means we can now configure it to use real AWS
services, rather than assuming we want to use the mock services.

The MOCK_AWS env var controls whether or not the mock AWS services are
deployed. If set to false, the channel adapter will be configured to use
real AWS services using the settings provided, otherwise these settings
will be ignored.

If real services are to be used, these must be set up prior to deployment.
They will not be provisioned by the deployment script.

* deployment: create channel adapter resources if requested

We now check the RDSS_ADAPTER_CREATE_AWS_RESOURCES env var to see if we are
supposed to create the required DynamoDB tables and Kinesis streams for the
channel adapter, if they don't already exist.

* deployment: check AWS CLI version is sufficient

We're now using 'dynamodb tag-resource' command, which was introduced in
v1.11.80 of the AWS CLI. This is newer than the default version available
on Ubuntu 14.04 LTS, for example, so requires the pip version be used
instead.

If the installed version of AWS CLI is not sufficient, the deploy script
will now stop will a fatal error.

* deployment: update teardown to include DynamoDB tables and Kinesis streams

We now teardown any DynamoDB tables that have been created. We also
optionally teardown the Kinesis streams too, but since these may be used by
other parts of the RDSS platform we do not do this by default. Instead, we
look to the DESTROY_KINESIS_QUEUES env var to see if we need to do so.

* RDSSARK-519 added role external id to deploy config

* RDSSARK-519 added storage service missing volume

* RDSSARK-519 remove publisher section since it is not used